fix(api): complete analyst extraction wiring and policy references
